# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang engineer test assessment - the Spy Cat Agency",
        "title": "Golang engineer test assessment - the Spy Cat Agency",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/mission/{mission_id}/target": {
            "post": {
                "description": "Add target to mission by mission_id and target_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "Add target to mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "mission_id's ID",
                        "name": "mission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Target"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/mission/{mission_id}/target/{target_id}": {
            "delete": {
                "description": "Delete target by mission id and target id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "Delete target",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "mission_id's ID",
                        "name": "mission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "target_id's ID",
                        "name": "target_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Update target's note  by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "Update target's note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "mission_id's ID",
                        "name": "mission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "target_id's ID",
                        "name": "target_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Target note",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateNotesPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.UpdateTargetNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/mission/{mission_id}/target_status/{target_id}": {
            "patch": {
                "description": "Update target's status  by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "Update target's status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "mission_id's ID",
                        "name": "mission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "target_id's ID",
                        "name": "target_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.UpdateTargetStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Target": {
            "type": "object",
            "required": [
                "complete",
                "country",
                "name",
                "notes"
            ],
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "notes": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "main.UpdateNotesPayload": {
            "type": "object",
            "required": [
                "notes"
            ],
            "properties": {
                "notes": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "store.UpdateTargetNote": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "mission_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "store.UpdateTargetStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "mission_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}